update generated controller
